consensus/entry_point_selector: add missing cast

+++ consensus/entry_point_selector/entry_point_selector.c
@@ -30,8 +30,8 @@ retcode_t iota_consensus_entry_point_selector_get_entry_point(
     entry_point_selector_t *const eps, size_t const depth,
     trit_array_p const ep) {
   retcode_t ret = RC_OK;
-  uint64_t milestone_index =
-      MAX(eps->mt->latest_solid_subtangle_milestone_index - depth - 1, 0);
+  uint64_t milestone_index = MAX(
+      (int64_t)eps->mt->latest_solid_subtangle_milestone_index - depth - 1, 0);
   iota_milestone_t milestone;
   iota_milestone_t *milestone_ptr = &milestone;
   iota_stor_pack_t pack = {.models = (void **)&milestone_ptr,
